Speed up parser even more
Simplifies (and thereby shrinks) the parse tree a lot more and resolves
most ambiguities in the grammar. There are still a couple left, but
they're much harder to fix (if doing so is possible at all). For now,
this will do - gives about a 40% performance boost on average, but
should give much more if emphasized text is involved (i.e. *bla*,
**bla**, ***bla***, etc.).

Contains a slight functional change in that an empty emphasized string
is no longer allowed (i.e. * * would no longer be accepted), but that
wouldn't have done anything to begin with (how are you going to
emphasize whitespace?), so not a problem.

Under #7