Allow aliases to be terminated explicitly
Same as commands, you can now terminate aliases explicitly with a
vertical bar. This is necessary if an alias is followed by something
other than whitespace. Take this example from the mockup project:

  @A met @B at a party in @HW|.

Without the explicit terminator, this would be parsed as

  @A| met @B| at a party in @HW.|

With it, we get the canonical parse result

  @A| met @B| at a party in @HW|.

instead, which is probably what was intended by the author.

Closes #1