Finish implementing the config framework
This adds tests, fixes the last few bugs in the framework and adds
support for loading default values when the config file is missing.

There's a 'breaking' change here that required adding a bunch of
aliases.zoia and dictionary.zoia files to the tests: we now require that
those two files exist (or their replacements if the config is present
and their paths have been altered). The docs have always made that clear
though, we just weren't enforcing it.

Closes #13