Implement alias evaluation
Also add a rudimentary example port of Shakespeare's Othello to the doc
folder. Useful for benchmarking and probably testing HTML generation in
the future.

This just kept growing and I had to keep fixing pylint/pytest failures.
Really should have made a branch and done it in a bunch of commits, then
squashed later on if necessary - oh well.

Closes #11
Under #18, some work on that got stuck here