One more round of speedups
Based on the realization that we don't even need the spaces in
textFragmentWord if we use the same trick as in lineElementsArg here.

Also dropped support for leading spaces in lineElementsInner, which
simplifies parser etc. a bit (just put your leading space before the
emphasized text, what will emphasizing some *spaces* even do?).