Restrict aliases to letters & numbers
This gives a huge usability improvement, because now you can write e.g.

 "@A!", @C shouted as she ran towards her friend.

and it would get parsed properly:

 "@A|!", @C| shouted as she ran towards her friend.

Before this change, you would have had to add the bar terminator to @A
to get this to work, otherwise it would have been parsed as:

 "@A!",| @C| shouted as she ran towards her friend.

Note that that would have changed slightly as soon as #22 is done, to
something like this:

 "@A!|", @C| shouted as she ran towards her friend.

But that also still needs a terminator to be correct. With how often
punctuation occurs after an alias (e.g. a character's name), that isn't
great. This way we only lose the ability to have aliases with
punctuation in them, which doesn't strike me as a problem. I did keep
numbers in because I can see those being useful (@A1, etc.).

Closes #39